requirements.txt
bcrypt==3.2.0
blinker==1.4
cffi==1.14.5
click==8.0.1
cryptography==3.4.7
dateutils==0.6.12
dnspython==1.16.0
email-validator==1.1.3
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-WTF==0.15.1
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mongoengine==0.23.1
mongoengine-goodjson==1.1.8
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.4
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
Werkzeug==2.0.1
WTForms==2.3.3
